H Teamara Black Coffee panetoimh kai se atomiko

+++ Team Black Coffee/Session-11/Libs/Employee.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Libs {
 
@@ -11,11 +8,11 @@ namespace Libs {
         public Guid ID { get; set; }
         public string? Name { get; set; }
         public string? Surname { get; set; }
-        public double Salary { get; set; }
-        
+        public decimal Salary { get; set; }
+
         public EmployeeType EmployeeType { get; set; }
 
-        
+
         public Employee() {
             ID = Guid.NewGuid();
         }
